[WIP] Fixed for MinGW
- Passed compilation processes
- Failed at linking libFoundation.dll

+++ Foundation/NSCFString.swift
@@ -207,7 +207,7 @@ internal func _CFSwiftStringFastContents(_ str: AnyObject) -> UnsafePointer<UniC
 }
 
 internal func _CFSwiftStringGetCString(_ str: AnyObject, buffer: UnsafeMutablePointer<Int8>, maxLength: Int, encoding: CFStringEncoding) -> Bool {
-    return (str as! NSString).getCString(buffer, maxLength: maxLength, encoding: CFStringConvertEncodingToNSStringEncoding(encoding))
+    return (str as! NSString).getCString(buffer, maxLength: maxLength, encoding: UInt(CFStringConvertEncodingToNSStringEncoding(encoding)))
 }
 
 internal func _CFSwiftStringIsUnicode(_ str: AnyObject) -> Bool {
